Inicio da importacao dos dados

+++ aterwebsrv/src/main/java/br/gov/df/emater/aterwebsrv/dao/funcional/LotacaoDao.java
@@ -9,7 +9,7 @@ import br.gov.df.emater.aterwebsrv.modelo.funcional.Emprego;
 import br.gov.df.emater.aterwebsrv.modelo.funcional.Lotacao;
 
 @Repository("LotacaoDao")
-public interface LotacaoDao extends JpaRepository<Lotacao, Long> {
+public interface LotacaoDao extends JpaRepository<Lotacao, Integer> {
 
 	List<Lotacao> findByEmprego(Emprego emprego);
